requirements.txt
discord.py
python-dotenv
pytz
requests
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
